Fix all imports, move errors to new subpackage, and make minor required changes to get tests to pass.
