+++ signac/core/synced_collections/__init__.py
@@ -1,0 +1,3 @@
+# Copyright (c) 2020 The Regents of the University of Michigan
+# All rights reserved.
+# This software is licensed under the BSD 3-Clause License.
